sk-FullProject1/backend/requirements.txt
asgiref==3.10.0
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
cffi==2.0.0
channels==4.3.1
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==4.2.26
dj-database-url==2.2.0
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
pillow==12.0.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyOpenSSL==25.3.0
service-identity==24.2.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
whitenoise==6.9.0
zope.interface==8.1.1
